+++ projects/ngx-forms-inputs/src/inputs/text-input/text-input.component.ts
@@ -1,7 +1,7 @@
 import {ControlContainer, FormsModule} from "@angular/forms";
 import {ChangeDetectionStrategy, ChangeDetectorRef, Component, Host, Optional, SkipSelf} from "@angular/core";
 import {harmonicaAnimation, IconDirective} from "@consensus-labs/ngx-tools";
-import {BaseInputComponent, FormScopeService} from "@consensus-labs/ngx-forms";
+import {BaseInputComponent, FormContext} from "@consensus-labs/ngx-forms";
 import {AsyncPipe, NgIf} from "@angular/common";
 import {MatLegacyTooltipModule} from "@angular/material/legacy-tooltip";
 import {MatLegacyInputModule} from "@angular/material/legacy-input";
@@ -29,7 +29,7 @@ export class TextInputComponent extends BaseInputComponent<string | undefined, s
   constructor(
     changes: ChangeDetectorRef,
     @Optional() @Host() @SkipSelf() controlContainer: ControlContainer,
-    @Optional() formScope: FormScopeService
+    @Optional() formScope: FormContext
   ) {
     super(changes, controlContainer, formScope);
   }