+++ projects/ngx-forms/src/lib/ngx-forms-tools/index.ts
@@ -2,4 +2,4 @@
 export * from './models/base-input-component';
 export * from './models/error-state-matchers';
 export * from './components/form-wrapper/form-wrapper.component';
-export * from './services/form-scope.service';
+export * from './services/form-context.service';