+++ projects/ngx-forms/src/lib/ngx-forms-tools/services/form-context.service.ts
@@ -1,0 +1,10 @@
+import {Injectable} from '@angular/core';
+import {BehaviorSubject} from "rxjs";
+
+@Injectable()
+export abstract class FormContext {
+
+  protected _readonly$ = new BehaviorSubject(false);
+  readonly$ = this._readonly$.asObservable();
+  get readonly() {return this._readonly$.value}
+}