+++ projects/ngx-forms/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @consensus-labs/ngx-forms
 
+## 15.2.0
+
+### Minor Changes
+
+- Change implementation of Form Context
+
 ## 15.1.3
 
 ### Patch Changes